refactor: :memo: renames poo submitted file name to match component name
