+++ packages/account/src/Components/poo/statuses/poo-submitted/index.ts
@@ -1,3 +1,3 @@
-import POOSubmitted from './submitted';
+import POOSubmitted from './poo-submitted';
 
 export default POOSubmitted;
